Move render-connection to lib
