+++ server/lib/render-connection.js
